+++ kubejs/client_scripts/jei_hide.js
@@ -1,0 +1,7 @@
+
+//This script hides things from JEI.
+
+JEIEvents.hideItems(event => {
+
+    event.hide('#jei:hidden_items')
+})