+++ kubejs/server_scripts/immersiveengineering.js
@@ -7,12 +7,26 @@ const ImmersiveEngineeringRecipes = (event) => {
     event.remove({output: 'immersiveengineering:nugget_steel'})
     event.remove({id: /^immersiveengineering:crafting\/nugget_.*_to_ingot_.*$/})
     event.remove({id: /^immersiveengineering:alloysmelter\/.*$/})
+    event.remove({output: 'immersiveengineering:ingot_nickel'})
+    event.remove({output: 'immersiveengineering:ingot_silver'})
+    event.remove({output: 'immersiveengineering:ingot_steel'})
+    event.remove({id: 'immersiveengineering:blastfurnace/steel_block'})
 
     //replace
     event.replaceInput(
         {id: 'immersiveengineering:crafting/voltmeter'},
         'minecraft:compass',
         '#minecraft:compasses'
+    )
+    event.replaceInput(
+        {id: 'tfc_ie_addon:metalpress/steel_block'},
+        '#forge:ingots/steel',
+        Item.of('tfc:metal/double_ingot/steel', 1)
+    )
+    event.replaceInput(
+        {id: 'tfc_ie_addon:metalpress/uranium_block'},
+        '#forge:ingots/uranium',
+        Item.of('tfc_ie_addon:metal/double_ingot/uranium', 1)
     )
 
     //TFC Heating