+++ kubejs/server_scripts/jei.js
@@ -1,0 +1,13 @@
+const JEItags = (event) => {
+
+    event.add('jei:hidden_items', ['minecraft:spawner', 'minecraft:gold_block', 'minecraft:gold_ingot',
+        'create:zinc_ingot', 'create:brass_ingot', 'dfc:metal/ingot/aluminum', 'dfc:metal/ingot/lead', 'dfc:metal/ingot/alumina',
+        'antiquelegacy:bronze_ingot', 'antiquelegacy:tin_ingot', 'epic_knights_ores_and_alloys:bronze_ingot', 'epic_knights_ores_and_alloys:tin_ingot',
+        'epic_knights_ores_and_alloys:steel_ingot', 'epic_knights_ores_and_alloys:silver_ingot', 'magistuarmory:steel_ingot', 'immersiveengineering:ingot_silver',
+        'immersiveengineering:ingot_nickel', 'immersiveengineering:ingot_steel', 'antiquelegacy:bronze_mixture', 'antiquelegacy:copper_nugget', 'antiquelegacy:raw_tin',
+        'antiquelegacy:tin_ore', 'antiquelegacy:deepslate_tin_ore', 'create:copper_nugget', 'epic_knights_ores_and_alloys:coppernugget', 'epic_knights_ores_and_alloys:bronzenugget',
+        'epic_knights_ores_and_alloys:tin_nugget', 'epic_knights_ores_and_alloys:silver_nugget', 'epic_knights_ores_and_alloys:steel_nugget', 'immersiveengineering:nugget_steel',
+        'immersiveengineering:alloybrick', 'immersiveengineering:slab_alloybrick', 'immersiveengineering:blastbrick', 'immersiveengineering:slab_blastbrick',
+        'immersiveengineering:slab_blastbrick_reinforced', 'immersiveengineering:blastbrick_reinforced', 'immersiveengineering:blast_furnace', 'immersiveengineering:alloy_smelter',
+        'immersiveengineering:advanced_blast_furnace'])
+}